requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LLM Providers
anthropic>=0.25.0
openai>=1.12.0
requests>=2.31.0

# GUI
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0

# Memory & Storage
sqlalchemy>=2.0.0
sqlite-vec>=0.0.1

# Tools & Utilities
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Observability
loguru>=0.7.0
prometheus-client>=0.19.0

# Vector Database
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Build & Packaging
pyinstaller>=6.3.0
pillow>=10.1.0

# Vision & Image Processing
pytesseract>=0.3.10
tiktoken>=0.5.0

# PDF Processing (Document Intelligence)
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Chat Integrations (Multi-Channel)
discord.py>=2.3.0
slack-sdk>=3.23.0

# Database Connectors (Zero-Copy)
psycopg2-binary>=2.9.0
supabase>=2.0.0

# Utilities
watchdog>=3.0.0
rich>=13.0.0

# Integration Dependencies
cloudflare>=2.11.0
stripe>=5.5.0
web3>=6.0.0
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
